requirements.txt
numpy~=1.26.4
opencv-python~=4.10.0.84
tqdm~=4.66.2
sahi~=0.11.23
Pillow~=10.1.0
tensorflow~=2.18.1